# docs/api/openapi-ms-usuarios.json
{"openapi":"3.1.0","info":{"title":"GreenBite - MS Usuarios API","description":"Microservicio de autenticación y gestión de usuarios. Persistencia con JPA/Hibernate sobre PostgreSQL (usuarios_db).","contact":{"name":"Equipo GreenBite","email":"equipo@greenbite.cl"},"license":{"name":"Uso académico - DUOC UC"},"version":"1.0.0"},"servers":[{"url":"http://localhost:4001","description":"Generated server url"}],"tags":[{"name":"Usuarios","description":"Autenticación y gestión de usuarios"}],"paths":{"/usuarios/{id}":{"get":{"tags":["Usuarios"],"summary":"Obtener usuario por id","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"put":{"tags":["Usuarios"],"summary":"Actualizar usuario","description":"Actualiza nombre, email y/o contraseña.","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"delete":{"tags":["Usuarios"],"summary":"Eliminar usuario","operationId":"remove","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/usuarios/register":{"post":{"tags":["Usuarios"],"summary":"Registrar un nuevo usuario","description":"Crea el usuario y devuelve sus datos junto con un token JWT.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/usuarios/login":{"post":{"tags":["Usuarios"],"summary":"Iniciar sesión","description":"Valida credenciales y devuelve los datos del usuario y un token JWT.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/usuarios":{"get":{"tags":["Usuarios"],"summary":"Listar usuarios","description":"Devuelve todos los usuarios ordenados por fecha de creación.","operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}}},"components":{"schemas":{"UpdateRequest":{"type":"object","properties":{"nombre":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"nombre":{"type":"string"},"email":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"type":"object","properties":{"nombre":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserResponse"},"token":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}}}}